Added part tests for actuators controller.

+++ modules/core/client/actuators/actuators-list/actuators.client.controller.js
@@ -1,33 +1,36 @@
 'use strict';
 
 angular.module('core')
-    .controller('ActuatorsController', ['$scope', 'actuatorsModel', 'actuatorsTypesModel', function ($scope, actuatorsModel, actuatorsTypesModel) {
+    .controller('ActuatorsController', function($scope, actuatorsModel, actuatorsTypesModel, confirmation) {
         $scope.actuatorsModel = actuatorsModel.model;
 
         $scope.actuatorsTypesModel = actuatorsTypesModel.model;
 
-        $scope.addEdit = function (controllerId) {
+        $scope.addEdit = function(controllerId) {
             actuatorsModel.addEdit(controllerId);
         };
 
-        $scope.delete = function (controllerId) {
-            actuatorsModel.delete(controllerId);
-        };
-
-        $scope.bulkDelete = actuatorsModel.bulkDelete;
-        $scope.bulkDeleteDisabled = function () {
-            return !actuatorsModel.model.actuators.some(function (controller) {
-                return controller.isSelected;
+        $scope.delete = function(controllerId) {
+            confirmation.confirm('Warning!', 'Do you really want to delete this item?', function() {
+                actuatorsModel.delete(controllerId);
             });
         };
 
-        $scope.toggleElectricitySwitcher = function(controller){
+        $scope.bulkDelete = function() {
+            confirmation.confirm('Warning!', 'Do you really want to delete these items?', function() {
+                actuatorsModel.bulkDelete();
+            });
+        };
+
+        $scope.bulkDeleteDisabled = actuatorsModel.bulkDeleteDisabled;
+
+        $scope.toggleElectricitySwitcher = function(controller) {
             actuatorsModel.changeValue(controller);
         };
 
-        $scope.onStopSlide = function($event, value, controller){
+        $scope.onStopSlide = function($event, value, controller) {
             actuatorsModel.changeValue(controller);
         };
 
-        actuatorsModel.load();
-    }]);
+        $scope.load = actuatorsModel.load;
+    });
